fix minor bug

+++ seeds/stock_seed.go
@@ -11,12 +11,12 @@ func SeedStocks(db *gorm.DB) {
 	// Seed 10 stock items for each of the first 10 items with random suppliers
 	for itemID := 1; itemID <= 10; itemID++ {
 		// Randomly select a supplier ID between 1 and 10 (adjust as needed)
-		supplierID := uint(rand.Intn(10) + 1)
+		supplierID := rand.Intn(10) + 1
 		initialStock := rand.Intn(100) + 1 // Random initial stock (1 to 100)
 
 		stock := models.Stock{
 			InitialStock: initialStock,
-			ItemID:       uint(itemID),
+			ItemID:       itemID,
 			SupplierID:   supplierID,
 		}
